add vsibench metric

+++ src/torchmetrics_ext/metrics/vqa/vsibench.py
@@ -47,7 +47,7 @@ class VSIBenchMetric(Metric):
     @staticmethod
     def _mean_relative_accuracy(pred, target, start, end, interval):
         num_pts = (end - start) / interval + 2
-        conf_intervals = np.linspace(start, end, int(num_pts), dtype=np.float32)
+        conf_intervals = np.linspace(start, end, int(num_pts), dtype=np.float64)
         accuracy = (abs(pred - target) / target) <= (1 - conf_intervals)
         return accuracy.mean()
